accounts/urls.py
from django.urls import path
from rest_framework.urlpatterns import format_suffix_patterns
from dj_rest_auth.views import LogoutView
from dj_rest_auth.registration.views import RegisterView
from rest_framework_simplejwt.views import TokenBlacklistView
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)
from .views import *

app_name = 'accounts'

urlpatterns = [
    path('register/', RegisterView.as_view(), name='register'),
    path('login/', CustomUserLogin.as_view(), name='login'),
    # path("logout/", LogoutView.as_view(), name="logout"),

    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('token/blacklist/', TokenBlacklistView.as_view(), name='token_blacklist'),
]

# urlpatterns = format_suffix_patterns(urlpatterns)
