requirements.txt
django==3.2
djangorestframework==3.14.0
dj-rest-auth
djangorestframework-simplejwt
django-cors-headers==4.1.0
django-sites
django-allauth
python-dotenv
